+++ vars/buildGoldenImage.groovy
@@ -53,7 +53,7 @@ def call(Map options) {
                                     ansiColor('xterm') {
                                         withAWS(credentials: 'os-imager-aws-creds', region: "${ec2_region}") {
                                             sh """
-                                            if [ ! -f bin/packer ]; then
+                                            if [ "\$(which packer)x" == "x" ] && [ ! -f bin/packer ]; then
                                                 mkdir -p bin
                                                 curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_linux_amd64.zip
                                                 curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_SHA256SUMS
@@ -103,7 +103,7 @@ def call(Map options) {
                                                 "ARTIFACTORY_URL=https://artifactory.saltstack.net/artifactory"
                                             ]) {
                                                 sh """
-                                                if [ ! -f bin/packer ]; then
+                                                if [ "\$(which packer)x" == "x" ] && [ ! -f bin/packer ]; then
                                                     mkdir -p bin
                                                     curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_linux_amd64.zip
                                                     curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_SHA256SUMS
@@ -294,7 +294,7 @@ def call(Map options) {
                                         checkout scm
                                         withAWS(credentials: 'os-imager-aws-creds', region: "${ec2_region}") {
                                             sh """
-                                            if [ ! -f bin/packer ]; then
+                                            if [ "\$(which packer)x" == "x" ] && [ ! -f bin/packer ]; then
                                                 mkdir -p bin
                                                 curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_linux_amd64.zip
                                                 curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_SHA256SUMS
@@ -457,7 +457,7 @@ def call(Map options) {
                                 checkout scm
                                 withAWS(credentials: 'os-imager-aws-creds', region: "${ec2_region}") {
                                     sh """
-                                    if [ ! -f bin/packer ]; then
+                                    if [ "\$(which packer)x" == "x" ] && [ ! -f bin/packer ]; then
                                         mkdir -p bin
                                         curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_linux_amd64.zip
                                         curl -O https://releases.hashicorp.com/packer/1.4.5/packer_1.4.5_SHA256SUMS
